type(fix): Fixed the texture bug for the hidden dealer card.
Biggest issues now are cards not displaying all the time so if the player goes bust it wont show the card that made them go bust. I'm going to PR because this is stable. Plus then I can review

Just noticed a weird bug, failed to load image errors in the console, images are loading so I think i have so left over code somewhere that I forgot to remove